Refined theory concentration
