+++ src/components/courses/DisplayCourseList.tsx
@@ -1,6 +1,12 @@
 import React from "react";
 import { Concentration } from "../../interfaces/concentration";
-import { AIConc, BioConc, SecurityConc, DataScienceConc, HPCConc, NetworksConc, TheoryConc } from "./CourseLists";
+import { AIConc } from "./concentrations/AIConc";
+import { BioConc } from "./concentrations/BioConc";
+import { DataScienceConc } from "./concentrations/DataScienceConc";
+import { HPCConc } from "./concentrations/HPCConc";
+import { NetworksConc } from "./concentrations/NetworksConc";
+import { SecurityConc } from "./concentrations/SecurityConc";
+import { TheoryConc } from "./concentrations/TheoryConc";
 import CONCENTRATIONS from "../../json/concentrations.json";
 import COURSES from "../../json/courses.json";
 import { Course as CourseType } from "../../interfaces/course";
